Fixed bugs and added functionality
Code in the item in the database is now the phase number
